requirements.txt
selenium
pandas
webdriver-manager
pyyaml
